deployment/requirements.txt
pyyaml
sagemaker
boto3